files/requirements.txt
# LiveKit Agents Core
livekit-agents>=0.10.0
livekit>=0.17.0

# Speech Recognition (STT)
livekit-plugins-deepgram>=0.8.0

# Language Model (LLM)
livekit-plugins-openai>=0.10.0

# Text-to-Speech (TTS)
livekit-plugins-cartesia>=0.3.0

# Voice Activity Detection
livekit-plugins-silero>=0.8.0

# Turn Detection
livekit-plugins-turn-detector>=0.1.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
websockets>=12.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
